requirements.txt
annotated-types==0.6.0
anyio==4.6.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloup==2.1.2
contourpy==1.2.0
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
distro==1.9.0
fonttools==4.47.2
git-dummy==0.1.2
git-sim==0.3.3
gitdb==4.0.11
GitPython==3.1.41
glcontext==2.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.6
isosurfaces==0.1.0
jiter==0.5.0
kiwisolver==1.4.5
lxml==5.3.0
manim==0.18.0
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
moderngl==5.9.0
moderngl-window==2.4.4
msedge-selenium-tools==3.141.4
multipledispatch==1.0.0
networkx==3.2.1
numpy==2.1.1
openai==1.50.0
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
Pillow==9.5.0
pybind11 @ file:///private/tmp/pybind11-20231211-91223-x6mu2j/pybind11-2.11.1
pycairo==1.27.0
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydub==0.25.1
pyglet==2.0.10
Pygments==2.17.2
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyparsing==3.1.1
pyrr==0.10.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
screeninfo==0.8.1
selenium==4.18.1
six==1.16.0
skia-pathops==0.7.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
svgelements==1.9.6
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.12.2
urllib3==2.1.0
watchdog==3.0.0
webdriver-manager==4.0.1
wsproto==1.2.0
